backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
requests==2.32.3
python-dotenv==1.1.0
pandas>=2.0.0
groq>=0.4.0
PyJWT>=2.8.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.9
